--- src/main/java/org/nextme/account_server/account/domain/entiry/Tran/TranType.java
@@ -1,6 +1,0 @@
-package org.nextme.account_server.account.domain.entiry.Tran;
-
-public enum TranType {
-    DEPOSIT,
-    WITHDRAW,
-}
