+++ src/main/java/org/nextme/account_server/account/domain/entity/Tran/TranType.java
@@ -1,0 +1,6 @@
+package org.nextme.account_server.account.domain.entity.Tran;
+
+public enum TranType {
+    DEPOSIT,
+    WITHDRAW,
+}
