+++ src/main/java/org/nextme/account_server/account/domain/repository/AccountRepository.java
@@ -1,11 +1,10 @@
 package org.nextme.account_server.account.domain.repository;
 
-import org.nextme.account_server.account.domain.entiry.Account;
-import org.nextme.account_server.account.domain.entiry.AccountId;
+import org.nextme.account_server.account.domain.entity.Account;
+import org.nextme.account_server.account.domain.entity.AccountId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
 import java.util.UUID;
 
 @Repository
