+++ src/main/java/org/nextme/account_server/account/domain/repository/BankRepository.java
@@ -1,6 +1,6 @@
 package org.nextme.account_server.account.domain.repository;
 
-import org.nextme.account_server.account.domain.entiry.Bank;
+import org.nextme.account_server.account.domain.entity.Bank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
